requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# Document processing
python-docx>=1.2.0
PyPDF2>=3.0.1
pdfplumber>=0.11.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation metrics
evaluate>=0.4.0
nltk>=3.8.0
rouge-score>=0.1.2
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional dependencies for better performance
# torch-audio  # For audio processing if needed
# sentencepiece  # For some tokenizers
# protobuf  # For some models

# Development dependencies (optional)
# jupyter
# ipykernel
# pytest
